requirements.txt
bidict==0.21.2
click==7.1.2
dnspython==1.16.0
eventlet==0.30.2
Flask==2.0.1
Flask-Caching==1.10.1
Flask-SocketIO==5.0.1
greenlet==1.1.0
gunicorn==20.1.0
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.18.0
python-engineio==4.1.0
python-socketio==5.2.1
selenium==3.141.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
